--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/domain/MutableCourseRepository.kt
@@ -1,7 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.domain
-
-interface MutableCourseRepository : CourseRepository {
-  fun clear()
-  fun saveCourse(courseEntity: CourseEntity)
-  fun saveAudiences(audiences: Set<Audience>)
-}
