+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/jparepo/CourseEntityRepository.kt
@@ -1,9 +1,9 @@
 package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.jparepo
 
-import org.springframework.data.repository.ListCrudRepository
+import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 import uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.domain.CourseEntity
 import java.util.UUID
 
 @Repository
-interface CourseEntityRepository : ListCrudRepository<CourseEntity, UUID>
+interface CourseEntityRepository : JpaRepository<CourseEntity, UUID>
